# Release v0.4.3.1b 2025-12-31
## New Features
1. make replication shard count agnostic (v0.4.3.0b 2025-12-28)
2. add debug replication mode (v0.4.3.0b 2025-12-28)

## Fixes
1. fix string push along with `CLIENT INFO` (v0.4.3.0b 2025-12-29)
2. hash test did not test hash (v0.4.3.0b 2025-12-29)
3. remove unused code (v0.4.3.1b 2025-12-30)
4. fix incr/decr for doubles

+++ src/keys.cpp
@@ -194,7 +194,7 @@ unsigned log_encoded_key(art::value_type key, bool start) {
     double dk;
     float fk;
     int64_t ik;
-    int64_t sk;
+    int32_t sk;
     const char *k;
     size_t kl;
     const unsigned char *enck = key.bytes;
